third_party/WebKit/LayoutTests/fast/regions/universal-selector-children-to-the-same-region-expected.txt

PASS: this is inline
PASS: and block content inside a region that is flowed into the same region

PASS: button
Bug 228566 - [CSS Regions] Elements in a region should be assignable to a named flow
